+++ NEWS.md
@@ -1,4 +1,8 @@
 # mFD (development version)
+* Add warning messages in `funct.dist` function: when using only continuous 
+traits, no weighting is realised for now.
+* Add informations in the General Tutorial and in the Continuous Traits,
+no weighting is realised for now.
 
 # mFD 1.0.3